requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.3
python-multipart==0.0.20
requests==2.32.3
httpx==0.28.1